refactor(capabilities): :recycle: make the command palette a generic design-system shell
The palette knew about chat, the easter egg hunt, the terminal and the site
search index, so it could never be part of a reusable design system. Invert it:
the design system now ships the shell plus two primitives, and the website
composes its own items.

- add CommandPaletteItem (owns the item styling that was copy-pasted in six
  files) and CommandPaletteGroup
- CommandPalette takes children, reports open state and query via callbacks, and
  drops the controlled selection so cmdk picks the first item itself
- close is now a window event (closeCommandPalette), symmetric with the existing
  openCommandPalette
- move the easter egg hunt, terminal, matrix rain and search result items to
  features/command-palette/site-command-palette
- move useSearch to features/hooks, terminal events to lib/terminal and the
  matrix rain panel event to lib/matrix-rain

Behaviour change: opening the palette with the keyboard shortcut now fires the
open tracking event too, which it previously did not.
