+++ src/components/design-system/organism/command-palette/command-palette-group/index.ts
@@ -1,0 +1,2 @@
+export { CommandPaletteGroup } from "./command-palette-group";
+export type { CommandPaletteGroupProps } from "./command-palette-group";
