+++ src/components/design-system/organism/command-palette/command-palette-item/index.ts
@@ -1,0 +1,2 @@
+export { CommandPaletteItem } from "./command-palette-item";
+export type { CommandPaletteItemProps } from "./command-palette-item";
