--- src/components/design-system/organism/command-palette/command-palette/easter-egg-hunt-item/index.ts
@@ -1,1 +1,0 @@
-export { EasterEggHuntItem } from "./easter-egg-hunt-item";
