+++ src/components/design-system/organism/command-palette/command-palette/index.ts
@@ -1,1 +1,2 @@
 export { CommandPalette } from "./command-palette";
+export type { CommandPaletteProps } from "./command-palette";
