--- src/components/design-system/organism/command-palette/command-palette/terminal-item/index.ts
@@ -1,1 +1,0 @@
-export { TerminalItem } from "./terminal-item";
