--- src/components/design-system/organism/command-palette/command-palette/toggle-motion-item/index.ts
@@ -1,1 +1,0 @@
-export { ToggleMotionItem } from "./toggle-motion-item";
