+++ src/components/design-system/organism/command-palette/toggle-motion-item/index.ts
@@ -1,0 +1,2 @@
+export { ToggleMotionItem } from "./toggle-motion-item";
+export type { ToggleMotionItemProps } from "./toggle-motion-item";
