+++ src/components/design-system/organism/command-palette/toggle-motion-item/toggle-motion-item.tsx
@@ -1,0 +1,31 @@
+"use client";
+
+import { TerminalLine } from "@/components/design-system/atoms/typography/terminal-blocks";
+import { CommandPaletteItem } from "@/components/design-system/organism/command-palette/command-palette-item";
+import { MdAnimation, MdDoDisturb } from "react-icons/md";
+import { FC } from "react";
+import { useToggleMotionItemStore } from "./use-toggle-motion-item-store";
+
+export interface ToggleMotionItemProps {
+    onTrack?: () => void;
+}
+
+export const ToggleMotionItem: FC<ToggleMotionItemProps> = ({ onTrack }) => {
+    const { state, effects } = useToggleMotionItemStore(onTrack);
+    const { motionEnabled } = state;
+    const { handleToggleMotion } = effects;
+
+    return (
+        <CommandPaletteItem value="toggle animations motion" onSelect={handleToggleMotion} closeOnSelect={false}>
+            <TerminalLine>
+                {motionEnabled ? (
+                    <MdDoDisturb className="mr-2 mb-0.5 inline" />
+                ) : (
+                    <MdAnimation className="mr-2 mb-0.5 inline" />
+                )}
+                {">"} Toggle Animations{" "}
+                <span className="text-accent/60 ml-1 font-mono text-xs">[{motionEnabled ? "ON" : "OFF"}]</span>
+            </TerminalLine>
+        </CommandPaletteItem>
+    );
+};
