+++ src/components/features/command-palette/site-command-palette/customize-matrix-rain-item/index.ts
@@ -1,0 +1,2 @@
+export { CustomizeMatrixRainItem } from "./customize-matrix-rain-item";
+export type { CustomizeMatrixRainItemProps } from "./customize-matrix-rain-item";
