+++ src/components/features/command-palette/site-command-palette/customize-matrix-rain-item/use-customize-matrix-rain-item-store.ts
@@ -1,8 +1,8 @@
 "use client";
 
-import { useWebGpuSupported } from "@/components/design-system/hooks/use-webgpu-supported";
 import { useReducedMotions } from "@/components/design-system/hooks/use-reduced-motions";
-import { openMatrixRainPanel } from "@/components/design-system/state/command-palette/command-palette-events";
+import { useWebGpuSupported } from "@/components/design-system/hooks/use-webgpu-supported";
+import { openMatrixRainPanel } from "@/lib/matrix-rain/matrix-rain-panel-events";
 import type { ComponentStore } from "@/types/component-store";
 import { useCallback } from "react";
 
@@ -15,16 +15,16 @@ interface CustomizeMatrixRainItemEffects {
 }
 
 export const useCustomizeMatrixRainItemStore = (
-    onClose: () => void,
+    onTrack?: () => void,
 ): ComponentStore<CustomizeMatrixRainItemState, CustomizeMatrixRainItemEffects> => {
     const webGpuSupported = useWebGpuSupported();
     const reducedMotion = useReducedMotions();
     const visible = webGpuSupported === true && !reducedMotion;
 
     const handleSelect = useCallback(() => {
-        onClose();
+        onTrack?.();
         openMatrixRainPanel();
-    }, [onClose]);
+    }, [onTrack]);
 
     return {
         state: { visible },
