+++ src/components/features/command-palette/site-command-palette/easter-egg-hunt-item/easter-egg-hunt-item.test.tsx
