+++ src/components/features/command-palette/site-command-palette/easter-egg-hunt-item/index.ts
@@ -1,0 +1,2 @@
+export { EasterEggHuntItem } from "./easter-egg-hunt-item";
+export type { EasterEggHuntItemProps } from "./easter-egg-hunt-item";
