+++ src/components/features/command-palette/site-command-palette/index.ts
@@ -1,0 +1,2 @@
+export { SiteCommandPalette } from "./site-command-palette";
+export type { SiteCommandPaletteProps } from "./site-command-palette";
