+++ src/components/features/command-palette/site-command-palette/search-result-item/index.ts
@@ -1,0 +1,2 @@
+export { SearchResultItem } from "./search-result-item";
+export type { SearchResultItemProps } from "./search-result-item";
