+++ src/components/features/command-palette/site-command-palette/search-result-item/search-result-item.tsx
@@ -1,14 +1,11 @@
 "use client";
 
-import { FC } from "react";
-import { Command } from "cmdk";
 import { TerminalListItem } from "@/components/design-system/molecules/terminal-list-item";
+import { CommandPaletteItem } from "@/components/design-system/organism/command-palette";
+import { FC } from "react";
 import { useSearchResultItemStore } from "./use-search-result-item-store";
 
-const ITEM_CLASS =
-    "px-4 py-2 cursor-pointer aria-selected:bg-accent-alpha-10 aria-selected:border-l-2 aria-selected:border-accent transition-colors duration-100";
-
-interface SearchResultItemProps {
+export interface SearchResultItemProps {
     title: string;
     description: string;
     slug: string;
@@ -20,8 +17,8 @@ export const SearchResultItem: FC<SearchResultItemProps> = ({ title, description
     const { handleSelect } = effects;
 
     return (
-        <Command.Item value={title} className={ITEM_CLASS} onSelect={handleSelect}>
+        <CommandPaletteItem value={title} onSelect={handleSelect}>
             <TerminalListItem title={title} description={description} />
-        </Command.Item>
+        </CommandPaletteItem>
     );
 };
