+++ src/components/features/command-palette/site-command-palette/search-result-item/use-search-result-item-store.ts
