+++ src/components/features/command-palette/site-command-palette/terminal-item/index.ts
@@ -1,0 +1,2 @@
+export { TerminalItem } from "./terminal-item";
+export type { TerminalItemProps } from "./terminal-item";
