+++ src/components/features/command-palette/site-command-palette/terminal-item/terminal-item.test.tsx
