+++ src/lib/matrix-rain/matrix-rain-panel-events.ts
@@ -1,0 +1,7 @@
+export const matrixRainPanelOpenEvent = "matrix-rain-panel-open";
+
+export const openMatrixRainPanel = () => {
+    if (typeof window !== "undefined") {
+        window.dispatchEvent(new Event(matrixRainPanelOpenEvent));
+    }
+};
