+++ src/lib/terminal/terminal-events.ts
